chore(infrastructure): rename namespace
